@@ -3,28 +3,26 @@ import React, { Component } from 'react';
 import Aux from '../../../hoc/Aux/Aux';
 import Button from '../../UI/Button/Button';
 
-class OrderSummary extends Component {
+const OrderSummary = props => {
 
-  render() {
-    const ingredientsSummary = Object.keys(this.props.ingredients)
+  const ingredientsSummary = Object.keys(props.ingredients)
     .map(igKey => {
-      return <li key={igKey}><span style={{textTransform: 'capitalize'}}>{igKey}</span>: {this.props.ingredients[igKey]}</li>
+      return <li key={igKey}><span style={{ textTransform: 'capitalize' }}>{igKey}</span>: {props.ingredients[igKey]}</li>
     });
-  
-    return (
-      <Aux>
-        <h3>Your Order</h3>
-        <p>A delicius burger with the following ingredients:</p>
-        <ul>
-          {ingredientsSummary}
-        </ul>
-        <p><strong>Total PRICE: {this.props.price.toFixed(2)}$</strong></p>
-        <p>Continue to Checkout?</p>
-        <Button btnType="Danger" clicked={this.props.purchaseCancelled}>CANCEL</Button>
-        <Button btnType="Success" clicked={this.props.purchaseContinued}>CONTINUE</Button>
-      </Aux>
-    );
-  }
+
+  return (
+    <Aux>
+      <h3>Your Order</h3>
+      <p>A delicius burger with the following ingredients:</p>
+      <ul>
+        {ingredientsSummary}
+      </ul>
+      <p><strong>Total PRICE: {props.price.toFixed(2)}$</strong></p>
+      <p>Continue to Checkout?</p>
+      <Button btnType="Danger" clicked={props.purchaseCancelled}>CANCEL</Button>
+      <Button btnType="Success" clicked={props.purchaseContinued}>CONTINUE</Button>
+    </Aux>
+  );
 }
 
 export default OrderSummary;
