@@ -1,32 +1,29 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 import classes from './Modal.module.css';
 import Aux from '../../../hoc/Aux/Aux';
 import Backdrop from '../Backdrop/Backdrop';
 
-class Modal extends Component {
+export const Modal = React.memo(props => {
 
-  shouldComponentUpdate(nextProps, nextState) {
-    return nextProps.show !== this.props.show || this.props.children !== nextProps.children;
-  }
+  // shouldComponentUpdate(nextProps, nextState) {
+  //   return nextProps.show !== this.props.show || this.props.children !== nextProps.children;
+  // }
 
-  render() {
 
-    return (
-      <Aux>
-        <Backdrop show={this.props.show} clicked={this.props.modalClosed} />
-        <div
-          className={classes.Modal}
-          style={{
-            transform: this.props.show ? 'translateY(0)' : 'translateY(100vh)',
-            opacity: this.props.show ? '1' : '0'
-          }}>
-          {this.props.children}
-        </div>
-      </Aux>
-    );
-  }
-}
+  return (
+    <Aux>
+      <Backdrop show={props.show} clicked={props.modalClosed} />
+      <div
+        className={classes.Modal}
+        style={{
+          transform: props.show ? 'translateY(0)' : 'translateY(100vh)',
+          opacity: props.show ? '1' : '0'
+        }}>
+        {props.children}
+      </div>
+    </Aux>
+  );
+}, (prevProps, nextProps) => nextProps.show === prevProps.show && prevProps.children === nextProps.children);
 
-export { Modal };
-export default Modal;
+export default React.memo(Modal, (prevProps, nextProps) => nextProps.show === prevProps.show && prevProps.children === nextProps.children);
