@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
@@ -6,131 +6,125 @@ import {
 	Button,
 	Input,
 	Spinner
- } from '../../components/UI/index';
+} from '../../components/UI/index';
 import classes from './Auth.module.css';
 import * as actions from '../../store/actions/index';
 import { updateObject, checkValidity } from '../../shared/utility';
 
-class Auth extends Component {
-	state = {
-		controls: {
-      email: {
-  	    elementType: 'input',
-        elementConfig: {
-          type: 'email',
-          placeholder: 'E-mail Adress'
-        },
-        value: '',
-        validation: {
-					required: true,
-					isEmail: true
-        },
-        valid: false,
-        touched: false
-      },
-      password: {
-  	    elementType: 'input',
-        elementConfig: {
-          type: 'password',
-          placeholder: 'Password'
-        },
-        value: '',
-        validation: {
-					required: true,
-					minLength: 6
-        },
-        valid: false,
-        touched: false
-      }
+const Auth = props => {
+	const [controls, setControls] = useState({
+		email: {
+			elementType: 'input',
+			elementConfig: {
+				type: 'email',
+				placeholder: 'E-mail Adress'
+			},
+			value: '',
+			validation: {
+				required: true,
+				isEmail: true
+			},
+			valid: false,
+			touched: false
 		},
-		isSignup: true
-	}
-
-	componentDidMount() {
-		if (!this.props.buildingBurger && this.props.authRedirectPath !== '/') {
-			this.props.onSetAuthRedirectPath();
+		password: {
+			elementType: 'input',
+			elementConfig: {
+				type: 'password',
+				placeholder: 'Password'
+			},
+			value: '',
+			validation: {
+				required: true,
+				minLength: 6
+			},
+			valid: false,
+			touched: false
 		}
-	}
+	});
 
+	const [isSignup, setIsSignup] = useState(true);
 
+	const { buildingBurger, authRedirectPath, onSetAuthRedirectPath } = props;
 
-  inputChangedHandler = (event, controlName) => {
-		const updatedControls = updateObject(this.state.controls, {
-			[controlName]: updateObject(this.state.controls[controlName], {
+	useEffect(() => {
+		if (!buildingBurger && authRedirectPath !== '/') {
+			onSetAuthRedirectPath();
+		}
+	}, [buildingBurger, authRedirectPath, onSetAuthRedirectPath]);
+
+	const inputChangedHandler = (event, controlName) => {
+		const updatedControls = updateObject(controls, {
+			[controlName]: updateObject(controls[controlName], {
 				value: event.target.value,
-				valid: checkValidity(event.target.value, this.state.controls[controlName].validation),
+				valid: checkValidity(event.target.value, controls[controlName].validation),
 				touched: true
 			})
 		});
-
-		this.setState({controls: updatedControls});
-  }
-	
-	submitHandler = (event) => {
-		event.preventDefault();
-
-		this.props.onAuth(this.state.controls.email.value, this.state.controls.password.value, this.state.isSignup);
+		setControls(updatedControls);
 	}
 
-	switchAuthModeHandler = () => {
-		this.setState(prevState => {
-			return{isSignup: !prevState.isSignup};
+	const submitHandler = (event) => {
+		event.preventDefault();
+
+		props.onAuth(controls.email.value, controls.password.value, isSignup);
+	}
+
+	const switchAuthModeHandler = () => {
+		setIsSignup(!isSignup);
+	}
+
+	const formElementsArray = [];
+	for (let key in controls) {
+		formElementsArray.push({
+			id: key,
+			config: controls[key]
 		})
 	}
 
-	render() {
-  	const formElementsArray = [];
-    for(let key in this.state.controls) {
-      formElementsArray.push({
-        id: key,
-        config: this.state.controls[key]
-      })
-		}
-		
-		let form = formElementsArray.map(formElement => (
-			<Input
-				key={formElement.id}
-				elementType={formElement.config.elementType}
-				elementConfig={formElement.config.elementConfig}
-				value={formElement.config.value}
-				invalid={!formElement.config.valid}
-				shouldValidate={formElement.config.validation}
-				touched={formElement.config.touched}
-				changed={(event) => this.inputChangedHandler(event, formElement.id)}
-				/>
-		));
+	let form = formElementsArray.map(formElement => (
+		<Input
+			key={formElement.id}
+			elementType={formElement.config.elementType}
+			elementConfig={formElement.config.elementConfig}
+			value={formElement.config.value}
+			invalid={!formElement.config.valid}
+			shouldValidate={formElement.config.validation}
+			touched={formElement.config.touched}
+			changed={(event) => inputChangedHandler(event, formElement.id)}
+		/>
+	));
 
-		if (this.props.loading) {
-			form = <Spinner />;
-		}
-
-		let errorMessge = null;
-
-		if(this.props.error) {
-			errorMessge = (<p style={{color: 'red'}}>{this.props.error.message}</p>);
-		}
-
-		let authRedirect = null;
-
-		if(this.props.isAuth) {
-			authRedirect = <Redirect to={this.props.authRedirectPath} />;
-		}
-
-		return (
-			<div className={classes.Auth}>
-				{authRedirect}
-				<h1>{this.state.isSignup ? "Signing up" : "Signing in"}</h1>
-				{errorMessge}
-				<form onSubmit={this.submitHandler}>
-					{form}
-					<Button btnType="Success">SUBMIT</Button>
-				</form>
-				<Button
-					clicked={this.switchAuthModeHandler}
-					btnType="Danger">SWITCH TO {this.state.isSignup ? "SIGNIN" : "SIGNUP"}</Button>
-			</div>
-		);
+	if (props.loading) {
+		form = <Spinner />;
 	}
+
+	let errorMessge = null;
+
+	if (props.error) {
+		errorMessge = (<p style={{ color: 'red' }}>{props.error.message}</p>);
+	}
+
+	let authRedirect = null;
+
+	if (props.isAuth) {
+		authRedirect = <Redirect to={props.authRedirectPath} />;
+	}
+
+	return (
+		<div className={classes.Auth}>
+			{authRedirect}
+			<h1>{isSignup ? "Signing up" : "Signing in"}</h1>
+			{errorMessge}
+			<form onSubmit={submitHandler}>
+				{form}
+				<Button btnType="Success">SUBMIT</Button>
+			</form>
+			<Button
+				clicked={switchAuthModeHandler}
+				btnType="Danger">SWITCH TO {isSignup ? "SIGNIN" : "SIGNUP"}</Button>
+		</div>
+	);
 }
 
 const mapStateToProps = state => {
