@@ -1,40 +1,39 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import CheckoutSummary from '../../components/Order/CheckoutSummary/CheckoutSummary';
 import ContactData from './ContactData/ContactData';
 
-class Checkout extends Component {
+const Checkout = props => {
 
-    checkoutCancelledHandler = () => {
-        this.props.history.goBack();
+    const checkoutCancelledHandler = () => {
+        props.history.goBack();
     }
 
-    checkoutContinuedHandler = () => {
-        this.props.history.replace('checkout/contact-data');
+    const checkoutContinuedHandler = () => {
+        props.history.replace('checkout/contact-data');
     }
 
-    render() {
-        let summary = <Redirect to="/" />;
-        const purchasedRedirect = this.props.purchased ? <Redirect to="/" /> : null;
 
-        if(this.props.ings) {
-            summary = (
-                <div>
-                    {purchasedRedirect}
-                    <CheckoutSummary
-                        ingredients={this.props.ings}
-                        checkoutCancelled={this.checkoutCancelledHandler}
-                        checkoutContinued={this.checkoutContinuedHandler} />
-                    <Route
-                        path={this.props.match.path + '/contact-data'}
-                        component={ContactData} />
-                </div>
-            );
-        }
-        return summary;
+    let summary = <Redirect to="/" />;
+    const purchasedRedirect = props.purchased ? <Redirect to="/" /> : null;
+
+    if (props.ings) {
+        summary = (
+            <div>
+                {purchasedRedirect}
+                <CheckoutSummary
+                    ingredients={props.ings}
+                    checkoutCancelled={checkoutCancelledHandler}
+                    checkoutContinued={checkoutContinuedHandler} />
+                <Route
+                    path={props.match.path + '/contact-data'}
+                    component={ContactData} />
+            </div>
+        );
     }
+    return summary;
 }
 
 const mapStateToProps = state => {
