@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 
 import Order from '../../components/Order/Order';
@@ -7,29 +7,29 @@ import withErrorHandler from '../../hoc/withErrorHandler/withErrorHandler';
 import Spinner from '../../components/UI/Spinner/Spinner';
 import * as actions from '../../store/actions/index';
 
-class Orders extends Component {
+const Orders = (props) => {
 
-  componentDidMount() {
-    this.props.onFetchOrders(this.props.token, this.props.userId);
-  }
+  const { onFetchOrders } = props;
 
-  render() {
-    let orders = <Spinner />;
-    if( !this.props.loading ) {
-      orders = this.props.orders.map(order => (
-        <Order
+  useEffect(() => {
+    onFetchOrders(props.token, props.userId);
+  }, [onFetchOrders]);
+
+  let orders = <Spinner />;
+  if (!props.loading) {
+    orders = props.orders.map(order => (
+      <Order
         key={order.id}
         ingredients={order.ingredients}
         price={order.price} />
-      ))
-    }
-
-    return (
-      <div>
-        {orders}
-      </div>
-    );
+    ))
   }
+
+  return (
+    <div>
+      {orders}
+    </div>
+  );
 }
 
 const mapStateToProps = state => {
