@@ -1,14 +1,13 @@
-import { delay } from 'redux-saga/effects';
-import { put } from 'redux-saga/effects';
+import { put, call, delay } from 'redux-saga/effects';
 import axios from 'axios';
 
-import * as actionTypes from '../actions/actionTypes'; 
+import * as actionTypes from '../actions/actionTypes';
 import * as actions from '../actions/index';
 
-export function*  logoutSaga(action) {
-	yield localStorage.removeItem('token');
-	yield localStorage.removeItem('expirationDate');
-	yield localStorage.removeItem('userId');
+export function* logoutSaga(action) {
+	yield call([localStorage, 'removeItem'], "token");
+	yield call([localStorage, 'removeItem'], "expirationDate");
+	yield call([localStorage, 'removeItem'], "userId");
 	yield put(actions.logoutSucceed());
 }
 
@@ -25,38 +24,38 @@ export function* authUserSaga(action) {
 		returnSecureToken: true
 	};
 	let url = "https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=AIzaSyCKTiaHaSC8g0ocrhukiNayxDYRenX2omQ";
-	if(!action.isSignup) {
+	if (!action.isSignup) {
 		url = "https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=AIzaSyCKTiaHaSC8g0ocrhukiNayxDYRenX2omQ";
 	}
 
 	try {
-	const res = yield axios.post(url, authData);
+		const res = yield axios.post(url, authData);
 
-	const expirationDate = yield new Date(new Date().getTime() + res.data.expiresIn * 1000);
-	yield localStorage.setItem('token', res.data.idToken);
-	yield localStorage.setItem('expirationDate', expirationDate);
-	yield localStorage.setItem('userId', res.data.localId);
-	yield put(actions.authSuccess(res.data.idToken, res.data.localId));
-	yield put(actions.checkAuthTimeout(res.data.expiresIn));
-	} catch(err) {
+		const expirationDate = yield new Date(new Date().getTime() + res.data.expiresIn * 1000);
+		yield localStorage.setItem('token', res.data.idToken);
+		yield localStorage.setItem('expirationDate', expirationDate);
+		yield localStorage.setItem('userId', res.data.localId);
+		yield put(actions.authSuccess(res.data.idToken, res.data.localId));
+		yield put(actions.checkAuthTimeout(res.data.expiresIn));
+	} catch (err) {
 		yield put(actions.authFail(err.response.data.error))
 	}
 }
 
 export function* authCheckStateSaga(action) {
-	
+
 	const token = yield localStorage.getItem('token');
-	if(!token) {
+	if (!token) {
 		yield put(actions.logout());
 	} else {
 		const expirationDate = yield new Date(localStorage.getItem("expirationDate"));
-			
-		if(expirationDate >= new Date()) {
+
+		if (expirationDate >= new Date()) {
 			const userId = yield localStorage.getItem('userId');
 			yield put(actions.authSuccess(token, userId));
-			yield put(actions.checkAuthTimeout((expirationDate.getTime() - new Date().getTime()) / 1000 ));
+			yield put(actions.checkAuthTimeout((expirationDate.getTime() - new Date().getTime()) / 1000));
 		} else {
-			 yield put(actions.logout());
+			yield put(actions.logout());
 		}
 	}
 }
